+++ tests/testthat/test_trees.R
@@ -42,3 +42,36 @@ test_that("is_tree works for directed out-trees", {
     expect_false(is_tree(g, mode="in"))
     expect_false(is_tree(g, mode="in", details=TRUE)$res)
 })
+
+test_that("the null graph is not a tree", {
+	expect_false(is_tree(make_empty_graph(0)))
+})
+
+test_that("a graph with a single vertex and no edges is tree", {
+	expect_true(is_tree(make_empty_graph(1)))
+})
+
+context("to_prufer, make_from_prufer")
+
+test_that("to_prufer and make_from_prufer works for trees", {
+    g <- make_tree(13, 3, mode="undirected")
+    seq <- to_prufer(g)
+    g2 <- make_from_prufer(seq)
+    expect_true(isomorphic(g, g2))
+
+    g <- make_tree(13, 3, mode="out")
+    seq <- to_prufer(g)
+    g2 <- make_from_prufer(seq)
+    g3 <- as.undirected(g)
+    expect_true(isomorphic(g2, g3))
+
+    g <- make_tree(13, 3, mode="in")
+    seq <- to_prufer(g)
+    g2 <- make_from_prufer(seq)
+    g3 <- as.undirected(g)
+    expect_true(isomorphic(g2, g3))
+})
+
+test_that("to_prufer prints an error for non-trees", {
+    expect_error(to_prufer(make_graph("zachary")), "must be a tree")
+})